feat: Add GitHub Actions workflow for Minimum Supported Rust
This commit adds a new GitHub Actions workflow file, `ci.yml`,
which is responsible for running tests and checks on the main
branch and pull requests. The workflow is triggered on pull
requests targeting the main branch and can also be manually
triggered using the workflow_dispatch event.

The workflow consists of a single job, `find-smallest-rust`,
which runs on the latest version of Ubuntu. It performs the
following steps:
1. Checks out the repository using the `actions/checkout` action.
2. Checks out the `NethermindEth/juno` repository using the `actions/checkout` action and stores it in the `juno` directory.
3. Uses a custom action, `./itself/action.yml`, to find the smallest supported Rust version in the `juno/core/rust/`, `juno/vm/rust/`, and `juno/starknet/rust/` directories.

This workflow is designed to ensure that the codebase is
compatible with the minimum supported Rust version specified
in the `juno` repository.